initial reorg, moved containers scss into components, added Base component, styled Spinner.js as example
