+++ app/components/Dataset.js
@@ -7,6 +7,7 @@ import ReactMarkdown from 'react-markdown'
 import { downloadDataset, deleteDataset, loadDatasetData } from '../actions/dataset'
 import { setQuery, runQuery, downloadQuery } from '../actions/query'
 
+import Base from './Base'
 import DatasetDataGrid from './DatasetDataGrid'
 import TabPanel from './TabPanel'
 import List from './List'
@@ -20,7 +21,7 @@ import MetadataEditorContainer from '../containers/MetadataEditor'
 
 import DatasetRefProps from '../propTypes/datasetRefProps'
 
-export default class Dataset extends React.Component {
+export default class Dataset extends Base {
   constructor (props) {
     super(props)
     this.state = {
@@ -268,7 +269,7 @@ export default class Dataset extends React.Component {
     }
   }
 
-  render () {
+  template (css) {
     const { datasetRef, readme } = this.props
     // const path = "/" + address.replace(".", "/", -1)
     // const hasData = (dataset && (dataset.url || dataset.file || dataset.data));
@@ -305,10 +306,6 @@ export default class Dataset extends React.Component {
 }
 
 Dataset.propTypes = {
-  // username.dataset address for this dataset, should
-  // be derived from url params
-  // path: PropTypes.string.isRequired,
-
   // the dataset model to display
   datasetRef: DatasetRefProps,
   // Readme model if available
@@ -322,16 +319,9 @@ Dataset.propTypes = {
   results: React.PropTypes.object,
   path: PropTypes.string,
 
-  // permissions stuff, will show things based on capabilities
-  // permissions: PropTypes.object.isRequired,
-
-  // action to load a dataset from passed-in address
-  // loadDatasetByAddress : PropTypes.func.isRequired,
-
   setQuery: PropTypes.func.isRequired,
   runQuery: PropTypes.func.isRequired,
   downloadDataset: PropTypes.func.isRequired
-  // loadDatasetReadme : PropTypes.func.isRequired
 }
 
 Dataset.defaultProps = {
