+++ app/scss/_components.scss
@@ -1,8 +1,12 @@
+@import "components/App";
+@import "components/ChartPicker";
 @import "components/CodeEditor";
+@import "components/CommandBar";
 @import "components/Console";
 @import "components/ContextPicker";
 @import "components/DatasetDataGrid";
 @import "components/DatasetItem";
+@import "components/Datasets";
 @import "components/DateInput";
 @import "components/DownloadResultsButton";
 @import "components/Fields";
@@ -11,16 +15,13 @@
 @import "components/Login";
 @import "components/MainMenu";
 @import "components/MonthCalendar";
+@import "components/Namespace";
 @import "components/Navbar";
+@import "components/Query";
 @import "components/QueryEditor";
 @import "components/QueryItem";
 @import "components/ResultsChart";
 @import "components/ResultsTable";
-@import "components/Signup";
-@import "components/Spinner";
-@import "components/SqlCode";
 @import "components/TabPanel";
-@import "components/TableColumnEditor";
-@import "components/TableColumnEditor";
 @import "components/View";
-@import "components/ChartPicker";
+@import "components/Welcome";
