--- app/scss/_containers.scss
@@ -1,8 +1,0 @@
-@import "containers/App";
-@import "containers/BetaSignup";
-@import "containers/Console";
-@import "containers/CommandBar";
-@import "containers/Namespace";
-@import "containers/Query";
-@import "containers/Welcome";
-@import "containers/Datasets";
