+++ app/scss/components/_Namespace.scss
@@ -1,0 +1,5 @@
+.namespace {
+	.label {
+		font-size : 20px;
+	}
+}
