--- app/scss/components/_Signup.scss
@@ -1,10 +1,0 @@
-#signup {
-	.form {
-		margin-top: 3em;
-		padding: 1em 1em 1.5em 1em;
-		.signup.btn {
-			width: 100%;
-			margin-bottom: 15px;
-		}
-	}
-}
