--- app/scss/components/_SqlCode.scss
@@ -1,5 +1,0 @@
-.sqlCode {
-  .load {
-    float: right;
-  }
-}
