+++ app/scss/components/_Welcome.scss
@@ -1,0 +1,3 @@
+#welcome {
+	
+}
