--- app/scss/containers/_App.scss
@@ -1,23 +1,0 @@
-.alert {
-  margin-bottom: 0;
-  position: fixed;
-  right: 0;
-  margin: 30px 20px;
-  overflow: auto;
-  z-index: 2;
-  min-width: 200px;
-  max-width: 40%;
-  padding: 10px 15px 10px 15px;
-	
-	.message {
-		float: left;
-		display: inline-block;
-		margin-left: 10px;
-		margin-right: 10px;
-	}
-	.dismiss {
-		float: right;
-		display: inline-block;
-		margin: 0 10px 0 0;
-	}
-}
