--- app/scss/containers/_BetaSignup.scss
@@ -1,20 +1,0 @@
-.betaSignup {
-	.close {
-		float : right;
-		background: none;
-		border: none;
-		color: $slate;
-		font-size: 1.5em;
-		margin-top: -0.8rem;
-	}
-	.title {
-		color: $red;
-		margin-top: 0.5rem;
-	}
-
-	.checkbox {
-		margin-left: 3%;
-		float: right;
-		margin-top: 5px;
-	}
-}
