--- app/scss/containers/_Datasets.scss
@@ -1,13 +1,0 @@
-.searchBox {
-    display: inline-block;
-    width: 50%;
-    font-size: 1rem;
-    line-height: 1.25;
-    color: #55595c;
-    background-color: #fff;
-    border: 0.5px solid rgba(0, 0, 0, 0.15);
-    overflow: auto;
-    border-radius: 0.25rem;
-    margin-bottom: 10px;
-    padding-left: 8px;
-}
