--- app/scss/containers/_Namespace.scss
@@ -1,5 +1,0 @@
-.namespace {
-	.label {
-		font-size : 20px;
-	}
-}
