--- app/scss/containers/_Welcome.scss
@@ -1,3 +1,0 @@
-#welcome {
-	
-}
