--- app/scss/containers/_console.scss
@@ -1,6 +1,0 @@
-#console {
-  .panel {
-    // padding-left: 20px;
-    // padding-right: 20px;
-  }
-}
