+++ app/scss/style.scss
@@ -15,6 +15,5 @@
 @import "modal";
 
 @import "components";
-@import "containers";
 
 @import "site";
